Add dummy cardano-crypto.js file
When using `cardano-addresses` as `source-repository-package` dependency in other projects, they fail to `cabal configure` for `ghcjs` because `jsbit/cardano-crypto.js` is missing.  This change introduces a dummy file that will need to be replaced.

It also moves the module for adding the emscripten compiled code into a separate file so we can use it more easily in other projects.